ai_backend/requirements.txt
Flask
langchain[groq]
langchain-huggingface
langchain-astradb
pypdf
inference_sdk
google-generativeai